Added Custom Conversation and Issue fixes
