+++ v2/spanner-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/SpannerToCassandraSourceDbIT.java
@@ -24,8 +24,7 @@ import com.datastax.oss.driver.api.core.cql.Row;
 import com.google.cloud.ByteArray;
 import com.google.cloud.Date;
 import com.google.cloud.Timestamp;
-import com.google.cloud.spanner.Mutation;
-import com.google.cloud.spanner.Value;
+import com.google.cloud.spanner.*;
 import com.google.cloud.teleport.metadata.SkipDirectRunnerTest;
 import com.google.cloud.teleport.metadata.TemplateIntegrationTest;
 import com.google.pubsub.v1.SubscriptionName;
@@ -33,18 +32,15 @@ import java.io.IOException;
 import java.math.BigDecimal;
 import java.nio.ByteBuffer;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import org.apache.beam.it.common.PipelineLauncher;
 import org.apache.beam.it.common.PipelineOperator;
 import org.apache.beam.it.common.utils.ResourceManagerUtils;
 import org.apache.beam.it.gcp.pubsub.PubsubResourceManager;
 import org.apache.beam.it.gcp.spanner.SpannerResourceManager;
 import org.apache.beam.it.gcp.storage.GcsResourceManager;
+import org.apache.beam.sdk.io.gcp.spanner.SpannerAccessor;
+import org.apache.beam.sdk.io.gcp.spanner.SpannerConfig;
 import org.apache.commons.codec.DecoderException;
 import org.apache.commons.codec.binary.Hex;
 import org.junit.AfterClass;
@@ -54,26 +50,31 @@ import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import org.junit.runners.model.MultipleFailureException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @Category({TemplateIntegrationTest.class, SkipDirectRunnerTest.class})
 @TemplateIntegrationTest(SpannerToSourceDb.class)
 @RunWith(JUnit4.class)
-public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbITBase {
+public class SpannerToCassandraSourceDbIT extends SpannerToSourceDbITBase {
+
+  private static final Logger LOG = LoggerFactory.getLogger(SpannerToCassandraSourceDbIT.class);
 
   private static final String SPANNER_DDL_RESOURCE =
-      "SpannerToCassandraSourceDbDatatypeIT/spanner-schema.sql";
+      "SpannerToCassandraSourceIT/spanner-schema.sql";
   private static final String CASSANDRA_SCHEMA_FILE_RESOURCE =
-      "SpannerToCassandraSourceDbDatatypeIT/cassandra-schema.sql";
+      "SpannerToCassandraSourceIT/cassandra-schema.sql";
   private static final String CASSANDRA_CONFIG_FILE_RESOURCE =
-      "SpannerToCassandraSourceDbDatatypeIT/cassandra-config-template.conf";
+      "SpannerToCassandraSourceIT/cassandra-config-template.conf";
 
-  private static final String TABLE = "AllDatatypeColumns";
-  private static final HashSet<SpannerToCassandraSourceDbDatatypeIT> testInstances =
-      new HashSet<>();
+  private static final String USER_TABLE = "Users";
+  private static final String ALL_DATA_TYPES_TABLE = "AllDatatypeColumns";
+  private static final String ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE = "AllDatatypeTransformation";
+  private static final HashSet<SpannerToCassandraSourceDbIT> testInstances = new HashSet<>();
   private static PipelineLauncher.LaunchInfo jobInfo;
   public static SpannerResourceManager spannerResourceManager;
   private static SpannerResourceManager spannerMetadataResourceManager;
-  public static CassandraSharedResourceManager cassandraResourceManager;
+  public static CassandraResourceManager cassandraResourceManager;
   private static GcsResourceManager gcsResourceManager;
   private static PubsubResourceManager pubsubResourceManager;
   private SubscriptionName subscriptionName;
@@ -87,7 +88,7 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
   @Before
   public void setUp() throws IOException {
     skipBaseCleanup = true;
-    synchronized (SpannerToCassandraSourceDbDatatypeIT.class) {
+    synchronized (SpannerToCassandraSourceDbIT.class) {
       testInstances.add(this);
       if (jobInfo == null) {
         spannerResourceManager = createSpannerDatabase(SPANNER_DDL_RESOURCE);
@@ -97,9 +98,9 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
         gcsResourceManager =
             GcsResourceManager.builder(artifactBucketName, getClass().getSimpleName(), credentials)
                 .build();
-        createCassandraSchema(cassandraResourceManager, CASSANDRA_SCHEMA_FILE_RESOURCE);
         createAndUploadCassandraConfigToGcs(
             gcsResourceManager, cassandraResourceManager, CASSANDRA_CONFIG_FILE_RESOURCE);
+        createCassandraSchema(cassandraResourceManager, CASSANDRA_SCHEMA_FILE_RESOURCE);
         pubsubResourceManager = setUpPubSubResourceManager();
         subscriptionName =
             createPubsubResources(
@@ -116,7 +117,8 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
                 null,
                 null,
                 null,
-                null);
+                null,
+                "cassandra");
       }
     }
   }
@@ -128,7 +130,7 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
    */
   @AfterClass
   public static void cleanUp() throws IOException {
-    for (SpannerToCassandraSourceDbDatatypeIT instance : testInstances) {
+    for (SpannerToCassandraSourceDbIT instance : testInstances) {
       instance.tearDownBase();
     }
     ResourceManagerUtils.cleanResources(
@@ -140,27 +142,117 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
   }
 
   @Test
-  public void spannerToCassandraSourceDataTypeConversionTest()
-      throws InterruptedException, IOException, MultipleFailureException {
+  public void spannerToCasandraSourceDbBasic() throws InterruptedException, IOException {
     assertThatPipeline(jobInfo).isRunning();
-    writeRowInSpanner();
-    assertRowInCassandraDB();
+    writeBasicRowInSpanner();
+    assertBasicRowInCassandraDB();
   }
 
-  private long getRowCount() {
-    String query = String.format("SELECT COUNT(*) FROM %s", TABLE);
+  @Test
+  public void spannerToCassandraSourceAllDataTypeConversionTest()
+      throws InterruptedException, IOException, MultipleFailureException {
+    assertThatPipeline(jobInfo).isRunning();
+    writeAllDataTypeRowsInSpanner();
+    assertAllDataTypeRowsRowInCassandraDB();
+  }
+
+  @Test
+  public void spannerToCassandraSourceDataTypeCustomConversionTest()
+      throws InterruptedException, IOException, MultipleFailureException {
+    assertThatPipeline(jobInfo).isRunning();
+    writeCustomRowsInSpanner();
+    assertCustomRowsInCassandraDB();
+  }
+
+  private long getRowCount(String tableName) {
+    String query = String.format("SELECT COUNT(*) FROM %s", tableName);
     ResultSet resultSet = cassandraResourceManager.executeStatement(query);
     Row row = resultSet.one();
     if (row != null) {
       return row.getLong(0);
     } else {
-      throw new RuntimeException("Query did not return a result for table: " + TABLE);
+      throw new RuntimeException("Query did not return a result for table: " + tableName);
     }
   }
 
-  private void writeRowInSpanner() {
+  private void writeBasicRowInSpanner() {
+    Mutation m1 =
+        Mutation.newInsertOrUpdateBuilder("users")
+            .set("id")
+            .to(1)
+            .set("full_name")
+            .to("A")
+            .set("from")
+            .to("B")
+            .build();
+    spannerResourceManager.write(m1);
+
+    Mutation m2 =
+        Mutation.newInsertOrUpdateBuilder("users2")
+            .set("id")
+            .to(2)
+            .set("full_name")
+            .to("BB")
+            .build();
+    spannerResourceManager.write(m2);
+
+    // Write a single record to Spanner for the given logical shard
+    // Add the record with the transaction tag as txBy=
+    SpannerConfig spannerConfig =
+        SpannerConfig.create()
+            .withProjectId(PROJECT)
+            .withInstanceId(spannerResourceManager.getInstanceId())
+            .withDatabaseId(spannerResourceManager.getDatabaseId());
+    SpannerAccessor spannerAccessor = SpannerAccessor.getOrCreate(spannerConfig);
+    spannerAccessor
+        .getDatabaseClient()
+        .readWriteTransaction(
+            Options.tag("txBy=forwardMigration"),
+            Options.priority(spannerConfig.getRpcPriority().get()))
+        .run(
+            (TransactionRunner.TransactionCallable<Void>)
+                transaction -> {
+                  Mutation m3 =
+                      Mutation.newInsertOrUpdateBuilder("users")
+                          .set("id")
+                          .to(3)
+                          .set("full_name")
+                          .to("GG")
+                          .set("from")
+                          .to("BB")
+                          .build();
+                  transaction.buffer(m3);
+                  return null;
+                });
+  }
+
+  private void assertBasicRowInCassandraDB() throws InterruptedException {
+    PipelineOperator.Result result =
+        pipelineOperator()
+            .waitForCondition(
+                createConfig(jobInfo, Duration.ofMinutes(10)), () -> getRowCount(USER_TABLE) == 1);
+    assertThatResult(result).meetsConditions();
+    Iterable<Row> rows;
+    try {
+      LOG.info("Reading from Cassandra table: {}", USER_TABLE);
+      rows = cassandraResourceManager.readTable(USER_TABLE);
+      LOG.info("Cassandra Rows: {}", rows.toString());
+    } catch (Exception e) {
+      throw new RuntimeException("Failed to read from Cassandra table: " + USER_TABLE, e);
+    }
+
+    assertThat(rows).hasSize(1);
+
+    Row row = rows.iterator().next();
+    LOG.info("Cassandra Row to Assert: {}", row.toString());
+    assertThat(row.getInt("id")).isEqualTo(1);
+    assertThat(row.getString("full_name")).isEqualTo("A");
+    assertThat(row.getString("from")).isEqualTo("B");
+  }
+
+  private void writeAllDataTypeRowsInSpanner() {
     Mutation mutation =
-        Mutation.newInsertOrUpdateBuilder(TABLE)
+        Mutation.newInsertOrUpdateBuilder(ALL_DATA_TYPES_TABLE)
             .set("varchar_column")
             .to("SampleVarchar")
             .set("tinyint_column")
@@ -245,6 +337,8 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
             .to(Value.json("[[\"apple\", \"banana\"], [\"carrot\", \"spinach\"]]"))
             .set("varint_column")
             .to(Value.bytes(ByteArray.copyFrom("b3f5ed4f".getBytes())))
+            .set("inet_column")
+            .to("192.168.1.10")
             .build();
 
     spannerResourceManager.write(mutation);
@@ -263,17 +357,19 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
     }
   }
 
-  private void assertRowInCassandraDB() throws InterruptedException, MultipleFailureException {
+  private void assertAllDataTypeRowsRowInCassandraDB()
+      throws InterruptedException, MultipleFailureException {
     PipelineOperator.Result result =
         pipelineOperator()
             .waitForCondition(
-                createConfig(jobInfo, Duration.ofMinutes(10)), () -> getRowCount() == 1);
+                createConfig(jobInfo, Duration.ofMinutes(10)),
+                () -> getRowCount(ALL_DATA_TYPES_TABLE) == 1);
     assertThatResult(result).meetsConditions();
     Iterable<Row> rows;
     try {
-      rows = cassandraResourceManager.readTable(TABLE);
+      rows = cassandraResourceManager.readTable(ALL_DATA_TYPES_TABLE);
     } catch (Exception e) {
-      throw new RuntimeException("Failed to read from Cassandra table: " + TABLE, e);
+      throw new RuntimeException("Failed to read from Cassandra table: " + ALL_DATA_TYPES_TABLE, e);
     }
 
     assertThat(rows).hasSize(1);
@@ -418,5 +514,236 @@ public class SpannerToCassandraSourceDbDatatypeIT extends SpannerToCassandraDbIT
         () ->
             assertThat(row.getBigInteger("varint_column"))
                 .isEqualTo(java.math.BigInteger.valueOf(7076111819049546854L)));
+  }
+
+  private void writeCustomRowsInSpanner() {
+    Mutation m;
+    m =
+        Mutation.newInsertOrUpdateBuilder(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE)
+            .set("varchar_column")
+            .to("SampleVarchar")
+            .set("tinyint_column")
+            .to(String.valueOf(127))
+            .set("text_column")
+            .to("This is some sample text data for the text column.")
+            .set("date_column")
+            .to(String.valueOf(Date.fromJavaUtilDate(java.sql.Date.valueOf("2025-01-27"))))
+            .set("smallint_column")
+            .to(String.valueOf(32767))
+            .set("mediumint_column")
+            .to(String.valueOf(8388607))
+            .set("int_column")
+            .to(String.valueOf(2147483647))
+            .set("bigint_column")
+            .to(String.valueOf(9223372036854775807L))
+            .set("float_column")
+            .to(String.valueOf(3.14159f))
+            .set("double_column")
+            .to(String.valueOf(2.718281828459045))
+            .set("decimal_column")
+            .to(new BigDecimal("12345.6789").toPlainString())
+            .set("datetime_column")
+            .to(String.valueOf(Timestamp.parseTimestamp("2025-01-27T10:30:00Z")))
+            .set("timestamp_column")
+            .to(String.valueOf(Timestamp.parseTimestamp("2025-01-27T10:30:00Z")))
+            .set("time_column")
+            .to("12:30:00")
+            .set("year_column")
+            .to("2025")
+            .set("char_column")
+            .to("CHAR_DATA")
+            .set("tinytext_column")
+            .to("Short text for tinytext.")
+            .set("mediumtext_column")
+            .to("Longer text data for mediumtext column.")
+            .set("longtext_column")
+            .to("Very long text data that exceeds the medium text column length for long text.")
+            .set("enum_column")
+            .to("OptionA")
+            .set("bool_column")
+            .to(String.valueOf(Boolean.TRUE))
+            .set("other_bool_column")
+            .to(String.valueOf(Boolean.FALSE))
+            .set("bytes_column")
+            .to(Value.bytes(ByteArray.copyFrom("SGVsbG8gd29ybGQ=".getBytes())))
+            .set("list_text_column")
+            .to(Value.json("[\"apple\", \"banana\", \"cherry\"]"))
+            .set("list_int_column")
+            .to(Value.json("[1, 2, 3, 4, 5]"))
+            .set("frozen_list_bigint_column")
+            .to(Value.json("[123456789012345, 987654321012345]"))
+            .set("set_text_column")
+            .to(Value.json("[\"apple\", \"orange\", \"banana\"]"))
+            .set("set_date_column")
+            .to(Value.json("[\"2025-01-27\", \"2025-02-01\"]"))
+            .set("frozen_set_bool_column")
+            .to(Value.json("[true, false]"))
+            .set("map_text_to_int_column")
+            .to(Value.json("{\"key1\": 10, \"key2\": 20}"))
+            .set("map_date_to_text_column")
+            .to(Value.json("{\"2025-01-27\": \"event1\", \"2025-02-01\": \"event2\"}"))
+            .set("frozen_map_int_to_bool_column")
+            .to(Value.json("{\"1\": true, \"2\": false}"))
+            .set("map_text_to_list_column")
+            .to(Value.json("{\"fruit\": [\"apple\", \"banana\"], \"color\": [\"red\", \"green\"]}"))
+            .set("map_text_to_set_column")
+            .to(
+                Value.json(
+                    "{\"fruit\": [\"apple\", \"banana\"], \"vegetables\": [\"carrot\", \"spinach\"]}"))
+            .set("set_of_maps_column")
+            .to(Value.json("[{\"key1\": 10, \"key2\": 20}, {\"keyA\": 5, \"keyB\": 10}]"))
+            .set("list_of_sets_column")
+            .to(Value.json("[[\"apple\", \"banana\"], [\"carrot\", \"spinach\"]]"))
+            .set("frozen_map_text_to_list_column")
+            .to(Value.json("{\"fruits\": [\"apple\", \"banana\"]}"))
+            .set("frozen_map_text_to_set_column")
+            .to(Value.json("{\"vegetables\": [\"carrot\", \"spinach\"]}"))
+            .set("frozen_set_of_maps_column")
+            .to(Value.json("[{\"key1\": 10, \"key2\": 20}, {\"keyA\": 5, \"keyB\": 10}]"))
+            .set("frozen_list_of_sets_column")
+            .to(Value.json("[[\"apple\", \"banana\"], [\"carrot\", \"spinach\"]]"))
+            .set("varint_column")
+            .to(Value.bytes(ByteArray.copyFrom("b3f5ed4f".getBytes())))
+            .build();
+
+    spannerResourceManager.write(m);
+
+    m =
+        Mutation.newUpdateBuilder(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE)
+            .set("varchar_column")
+            .to("SampleVarchar")
+            .set("tinyint_column")
+            .to(String.valueOf(122))
+            .set("text_column")
+            .to("This is some sample text data for the text column.")
+            .set("date_column")
+            .to(String.valueOf(Date.fromJavaUtilDate(java.sql.Date.valueOf("2025-01-27"))))
+            .set("smallint_column")
+            .to(String.valueOf(32767))
+            .set("mediumint_column")
+            .to(String.valueOf(8388607))
+            .set("int_column")
+            .to(String.valueOf(2147483647))
+            .set("bigint_column")
+            .to(String.valueOf(9223372036854775807L))
+            .set("float_column")
+            .to(String.valueOf(3.14159f))
+            .set("double_column")
+            .to(String.valueOf(2.718281828459045))
+            .set("decimal_column")
+            .to(new BigDecimal("12345.6789").toPlainString())
+            .set("datetime_column")
+            .to(String.valueOf(Timestamp.parseTimestamp("2025-01-27T10:30:00Z")))
+            .set("timestamp_column")
+            .to(String.valueOf(Timestamp.parseTimestamp("2025-01-27T10:30:00Z")))
+            .set("time_column")
+            .to("12:30:00")
+            .set("year_column")
+            .to("2025")
+            .set("char_column")
+            .to("CHAR_DATA")
+            .set("tinytext_column")
+            .to("Short text for tinytext.")
+            .set("mediumtext_column")
+            .to("Longer text data for mediumtext column.")
+            .set("longtext_column")
+            .to("Very long text data that exceeds the medium text column length for long text.")
+            .set("enum_column")
+            .to("OptionA")
+            .set("bool_column")
+            .to(String.valueOf(Boolean.TRUE))
+            .set("other_bool_column")
+            .to(String.valueOf(Boolean.FALSE))
+            .set("bytes_column")
+            .to(Value.bytes(ByteArray.copyFrom("SGVsbG8gd29ybGQ=".getBytes())))
+            .set("list_text_column")
+            .to(Value.json("[\"apple\", \"banana\", \"cherry\"]"))
+            .set("list_int_column")
+            .to(Value.json("[1, 2, 3, 4, 5]"))
+            .set("frozen_list_bigint_column")
+            .to(Value.json("[123456789012345, 987654321012345]"))
+            .set("set_text_column")
+            .to(Value.json("[\"apple\", \"orange\", \"banana\"]"))
+            .set("set_date_column")
+            .to(Value.json("[\"2025-01-27\", \"2025-02-01\"]"))
+            .set("frozen_set_bool_column")
+            .to(Value.json("[true, false]"))
+            .set("map_text_to_int_column")
+            .to(Value.json("{\"key1\": 10, \"key2\": 20}"))
+            .set("map_date_to_text_column")
+            .to(Value.json("{\"2025-01-27\": \"event1\", \"2025-02-01\": \"event2\"}"))
+            .set("frozen_map_int_to_bool_column")
+            .to(Value.json("{\"1\": true, \"2\": false}"))
+            .set("map_text_to_list_column")
+            .to(Value.json("{\"fruit\": [\"apple\", \"banana\"], \"color\": [\"red\", \"green\"]}"))
+            .set("map_text_to_set_column")
+            .to(
+                Value.json(
+                    "{\"fruit\": [\"apple\", \"banana\"], \"vegetables\": [\"carrot\", \"spinach\"]}"))
+            .set("set_of_maps_column")
+            .to(Value.json("[{\"key1\": 10, \"key2\": 20}, {\"keyA\": 5, \"keyB\": 10}]"))
+            .set("list_of_sets_column")
+            .to(Value.json("[[\"apple\", \"banana\"], [\"carrot\", \"spinach\"]]"))
+            .set("frozen_map_text_to_list_column")
+            .to(Value.json("{\"fruits\": [\"apple\", \"banana\"]}"))
+            .set("frozen_map_text_to_set_column")
+            .to(Value.json("{\"vegetables\": [\"carrot\", \"spinach\"]}"))
+            .set("frozen_set_of_maps_column")
+            .to(Value.json("[{\"key1\": 10, \"key2\": 20}, {\"keyA\": 5, \"keyB\": 10}]"))
+            .set("frozen_list_of_sets_column")
+            .to(Value.json("[[\"apple\", \"banana\"], [\"carrot\", \"spinach\"]]"))
+            .set("varint_column")
+            .to(Value.bytes(ByteArray.copyFrom("b3f5ed4f".getBytes())))
+            .build();
+
+    spannerResourceManager.write(m);
+    m = Mutation.delete(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE, Key.of("SampleVarchar"));
+    spannerResourceManager.write(m);
+  }
+
+  private void assertCustomRowsInCassandraDB() throws MultipleFailureException {
+    PipelineOperator.Result result =
+        pipelineOperator()
+            .waitForCondition(
+                createConfig(jobInfo, Duration.ofMinutes(10)),
+                () -> getRowCount(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE) == 1);
+    assertThatResult(result).meetsConditions();
+
+    Iterable<Row> rows;
+    try {
+      rows = cassandraResourceManager.readTable(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE);
+    } catch (Exception e) {
+      throw new RuntimeException(
+          "Failed to read from Cassandra table: " + ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE, e);
+    }
+
+    assertThat(rows).hasSize(1);
+    Row row = rows.iterator().next();
+    assertThat(rows).hasSize(1);
+
+    assertAll(
+        () -> assertThat(row.getString("varchar_column")).isEqualTo("SampleVarchar"),
+        () -> assertThat(row.getByte("tinyint_column")).isEqualTo((byte) 127));
+
+    ResultSet resultSets =
+        cassandraResourceManager.executeStatement(
+            String.format("SELECT * FROM %s", ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE));
+    resultSets.forEach(
+        resultSet -> {
+          try {
+            assertAll(
+                () -> assertThat(resultSet.getString("varchar_column")).isEqualTo("SampleVarchar"),
+                () -> assertThat(resultSet.getByte("tinyint_column")).isEqualTo((byte) 122));
+          } catch (MultipleFailureException e) {
+            throw new RuntimeException(e);
+          }
+        });
+
+    PipelineOperator.Result result1 =
+        pipelineOperator()
+            .waitForCondition(
+                createConfig(jobInfo, Duration.ofMinutes(25)),
+                () -> getRowCount(ALL_DATA_TYPES_CUSTOM_CONVERSION_TABLE) == 0);
+    assertThatResult(result1).meetsConditions();
   }
 }
