--- v2/spanner-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/SpannerToSourceDbCassandraIT.java
@@ -1,226 +1,0 @@
-/*
- * Copyright (C) 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.teleport.v2.templates;
-
-import static com.google.common.truth.Truth.assertThat;
-import static org.apache.beam.it.truthmatchers.PipelineAsserts.assertThatPipeline;
-import static org.apache.beam.it.truthmatchers.PipelineAsserts.assertThatResult;
-
-import com.datastax.oss.driver.api.core.cql.ResultSet;
-import com.datastax.oss.driver.api.core.cql.Row;
-import com.google.cloud.spanner.Mutation;
-import com.google.cloud.spanner.Options;
-import com.google.cloud.spanner.TransactionRunner;
-import com.google.cloud.teleport.metadata.SkipDirectRunnerTest;
-import com.google.cloud.teleport.metadata.TemplateIntegrationTest;
-import com.google.pubsub.v1.SubscriptionName;
-import java.io.IOException;
-import java.time.Duration;
-import java.util.HashSet;
-import org.apache.beam.it.common.PipelineLauncher;
-import org.apache.beam.it.common.PipelineOperator;
-import org.apache.beam.it.common.utils.ResourceManagerUtils;
-import org.apache.beam.it.gcp.pubsub.PubsubResourceManager;
-import org.apache.beam.it.gcp.spanner.SpannerResourceManager;
-import org.apache.beam.it.gcp.storage.GcsResourceManager;
-import org.apache.beam.sdk.io.gcp.spanner.SpannerAccessor;
-import org.apache.beam.sdk.io.gcp.spanner.SpannerConfig;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Category({TemplateIntegrationTest.class, SkipDirectRunnerTest.class})
-@TemplateIntegrationTest(SpannerToSourceDb.class)
-@RunWith(JUnit4.class)
-public class SpannerToSourceDbCassandraIT extends SpannerToCassandraDbITBase {
-
-  private static final Logger LOG = LoggerFactory.getLogger(SpannerToSourceDbCassandraIT.class);
-
-  private static final String SPANNER_DDL_RESOURCE =
-      "SpannerToCassandraSourceIT/spanner-schema.sql";
-  private static final String CASSANDRA_SCHEMA_FILE_RESOURCE =
-      "SpannerToCassandraSourceIT/cassandra-schema.sql";
-  private static final String CASSANDRA_CONFIG_FILE_RESOURCE =
-      "SpannerToCassandraSourceIT/cassandra-config-template.conf";
-
-  private static final String TABLE = "Users";
-  private static final HashSet<SpannerToSourceDbCassandraIT> testInstances = new HashSet<>();
-  private static PipelineLauncher.LaunchInfo jobInfo;
-  public static SpannerResourceManager spannerResourceManager;
-  private static SpannerResourceManager spannerMetadataResourceManager;
-  public static CassandraSharedResourceManager cassandraResourceManager;
-  private static GcsResourceManager gcsResourceManager;
-  private static PubsubResourceManager pubsubResourceManager;
-  private SubscriptionName subscriptionName;
-
-  /**
-   * Setup resource managers and Launch dataflow job once during the execution of this test class.
-   *
-   * @throws IOException
-   */
-  @Before
-  public void setUp() throws IOException {
-    skipBaseCleanup = true;
-    synchronized (SpannerToSourceDbCassandraIT.class) {
-      testInstances.add(this);
-      if (jobInfo == null) {
-        spannerResourceManager = createSpannerDatabase(SPANNER_DDL_RESOURCE);
-        spannerMetadataResourceManager = createSpannerMetadataDatabase();
-
-        cassandraResourceManager = generateKeyspaceAndBuildCassandraResource();
-        gcsResourceManager =
-            GcsResourceManager.builder(artifactBucketName, getClass().getSimpleName(), credentials)
-                .build();
-        createAndUploadCassandraConfigToGcs(
-            gcsResourceManager, cassandraResourceManager, CASSANDRA_CONFIG_FILE_RESOURCE);
-        createCassandraSchema(cassandraResourceManager, CASSANDRA_SCHEMA_FILE_RESOURCE);
-        pubsubResourceManager = setUpPubSubResourceManager();
-        subscriptionName =
-            createPubsubResources(
-                getClass().getSimpleName(),
-                pubsubResourceManager,
-                getGcsPath("dlq", gcsResourceManager).replace("gs://" + artifactBucketName, ""));
-        jobInfo =
-            launchDataflowJob(
-                gcsResourceManager,
-                spannerResourceManager,
-                spannerMetadataResourceManager,
-                subscriptionName.toString(),
-                null,
-                null,
-                null,
-                null,
-                null);
-      }
-    }
-  }
-
-  /**
-   * Cleanup dataflow job and all the resources and resource managers.
-   *
-   * @throws IOException
-   */
-  @AfterClass
-  public static void cleanUp() throws IOException {
-    for (SpannerToSourceDbCassandraIT instance : testInstances) {
-      instance.tearDownBase();
-    }
-    ResourceManagerUtils.cleanResources(
-        spannerResourceManager,
-        cassandraResourceManager,
-        spannerMetadataResourceManager,
-        gcsResourceManager,
-        pubsubResourceManager);
-  }
-
-  @Test
-  public void spannerToCasandraSourceDbBasic() throws InterruptedException, IOException {
-    assertThatPipeline(jobInfo).isRunning();
-    writeRowInSpanner();
-    assertRowInCassandraDB();
-  }
-
-  private long getRowCount() {
-    String query = String.format("SELECT COUNT(*) FROM %s", TABLE);
-    ResultSet resultSet = cassandraResourceManager.executeStatement(query);
-    Row row = resultSet.one();
-    if (row != null) {
-      return row.getLong(0);
-    } else {
-      throw new RuntimeException("Query did not return a result for table: " + TABLE);
-    }
-  }
-
-  private void writeRowInSpanner() {
-    Mutation m1 =
-        Mutation.newInsertOrUpdateBuilder("users")
-            .set("id")
-            .to(1)
-            .set("full_name")
-            .to("A")
-            .set("from")
-            .to("B")
-            .build();
-    spannerResourceManager.write(m1);
-
-    Mutation m2 =
-        Mutation.newInsertOrUpdateBuilder("users2")
-            .set("id")
-            .to(2)
-            .set("full_name")
-            .to("BB")
-            .build();
-    spannerResourceManager.write(m2);
-
-    // Write a single record to Spanner for the given logical shard
-    // Add the record with the transaction tag as txBy=
-    SpannerConfig spannerConfig =
-        SpannerConfig.create()
-            .withProjectId(PROJECT)
-            .withInstanceId(spannerResourceManager.getInstanceId())
-            .withDatabaseId(spannerResourceManager.getDatabaseId());
-    SpannerAccessor spannerAccessor = SpannerAccessor.getOrCreate(spannerConfig);
-    spannerAccessor
-        .getDatabaseClient()
-        .readWriteTransaction(
-            Options.tag("txBy=forwardMigration"),
-            Options.priority(spannerConfig.getRpcPriority().get()))
-        .run(
-            (TransactionRunner.TransactionCallable<Void>)
-                transaction -> {
-                  Mutation m3 =
-                      Mutation.newInsertOrUpdateBuilder("users")
-                          .set("id")
-                          .to(3)
-                          .set("full_name")
-                          .to("GG")
-                          .set("from")
-                          .to("BB")
-                          .build();
-                  transaction.buffer(m3);
-                  return null;
-                });
-  }
-
-  private void assertRowInCassandraDB() throws InterruptedException {
-    PipelineOperator.Result result =
-        pipelineOperator()
-            .waitForCondition(
-                createConfig(jobInfo, Duration.ofMinutes(10)), () -> getRowCount() == 1);
-    assertThatResult(result).meetsConditions();
-    Iterable<Row> rows;
-    try {
-      LOG.info("Reading from Cassandra table: {}", TABLE);
-      rows = cassandraResourceManager.readTable(TABLE);
-      LOG.info("Cassandra Rows: {}", rows.toString());
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to read from Cassandra table: " + TABLE, e);
-    }
-
-    assertThat(rows).hasSize(1);
-
-    Row row = rows.iterator().next();
-    LOG.info("Cassandra Row to Assert: {}", row.toString());
-    assertThat(row.getInt("id")).isEqualTo(1);
-    assertThat(row.getString("full_name")).isEqualTo("A");
-    assertThat(row.getString("from")).isEqualTo("B");
-  }
-}
