+++ v2/spanner-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/SpannerToSourceDbITBase.java
@@ -25,14 +25,18 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.pubsub.v1.SubscriptionName;
 import com.google.pubsub.v1.TopicName;
-import java.io.IOException;
+import java.io.*;
 import java.nio.charset.StandardCharsets;
+import java.time.format.DateTimeFormatter;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import org.apache.beam.it.common.PipelineLauncher;
 import org.apache.beam.it.common.utils.IORedirectUtil;
 import org.apache.beam.it.common.utils.PipelineUtils;
+import org.apache.beam.it.common.utils.ResourceManagerUtils;
 import org.apache.beam.it.gcp.TemplateTestBase;
 import org.apache.beam.it.gcp.artifacts.utils.ArtifactUtils;
 import org.apache.beam.it.gcp.pubsub.PubsubResourceManager;
@@ -53,14 +57,27 @@ public abstract class SpannerToSourceDbITBase extends TemplateTestBase {
         SpannerResourceManager.builder("rr-main-" + testName, PROJECT, REGION)
             .maybeUseStaticInstance()
             .build();
-    String ddl =
-        String.join(
-            " ",
-            Resources.readLines(Resources.getResource(spannerSchemaFile), StandardCharsets.UTF_8));
-    ddl = ddl.trim();
-    String[] ddls = ddl.split(";");
+
+    String ddl;
+    try (InputStream inputStream =
+        Thread.currentThread().getContextClassLoader().getResourceAsStream(spannerSchemaFile)) {
+      if (inputStream == null) {
+        throw new FileNotFoundException("Resource file not found: " + spannerSchemaFile);
+      }
+      try (BufferedReader reader =
+          new BufferedReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))) {
+        ddl = reader.lines().collect(Collectors.joining("\n"));
+      }
+    }
+
+    if (ddl.isBlank()) {
+      throw new IllegalStateException("DDL file is empty: " + spannerSchemaFile);
+    }
+
+    String[] ddls = ddl.trim().split(";");
     for (String d : ddls) {
-      if (!d.isBlank()) {
+      d = d.trim();
+      if (!d.isEmpty()) {
         spannerResourceManager.executeDdlStatement(d);
       }
     }
@@ -116,6 +133,68 @@ public abstract class SpannerToSourceDbITBase extends TemplateTestBase {
     gcsResourceManager.createArtifact("input/shard.json", shardFileContents);
   }
 
+  protected CassandraResourceManager generateKeyspaceAndBuildCassandraResource() {
+    String keyspaceName =
+        ResourceManagerUtils.generateResourceId(
+                testName,
+                Pattern.compile("[/\\\\. \"\u0000$]"),
+                "-",
+                27,
+                DateTimeFormatter.ofPattern("yyyyMMdd-HHmmss-SSSSSS"))
+            .replace('-', '_');
+    if (keyspaceName.length() > 48) {
+      keyspaceName = keyspaceName.substring(0, 48);
+    }
+    return CassandraResourceManager.builder(testName)
+        .setKeyspaceName(keyspaceName)
+        .sePreGeneratedKeyspaceName(true)
+        .build();
+  }
+
+  protected void createCassandraSchema(
+      CassandraResourceManager cassandraResourceManager, String cassandraSchemaFile)
+      throws IOException {
+    String ddl =
+        String.join(
+            " ",
+            Resources.readLines(
+                Resources.getResource(cassandraSchemaFile), StandardCharsets.UTF_8));
+    ddl = ddl.trim();
+    String[] ddls = ddl.split(";");
+    for (String d : ddls) {
+      if (!d.isBlank()) {
+        cassandraResourceManager.execute(d);
+      }
+    }
+  }
+
+  public void createAndUploadCassandraConfigToGcs(
+      GcsResourceManager gcsResourceManager,
+      CassandraResourceManager cassandraResourceManagers,
+      String cassandraConfigFile)
+      throws IOException {
+
+    String host = cassandraResourceManagers.getHost();
+    int port = cassandraResourceManagers.getPort();
+    String keyspaceName = cassandraResourceManagers.getKeyspaceName();
+    String cassandraConfigContents;
+    try (InputStream inputStream =
+        Thread.currentThread().getContextClassLoader().getResourceAsStream(cassandraConfigFile)) {
+      if (inputStream == null) {
+        throw new FileNotFoundException("Resource file not found: " + cassandraConfigFile);
+      }
+      cassandraConfigContents = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+    }
+
+    cassandraConfigContents =
+        cassandraConfigContents
+            .replace("##host##", host)
+            .replace("##port##", Integer.toString(port))
+            .replace("##keyspace##", keyspaceName);
+
+    gcsResourceManager.createArtifact("input/cassandra-config.conf", cassandraConfigContents);
+  }
+
   public PipelineLauncher.LaunchInfo launchDataflowJob(
       GcsResourceManager gcsResourceManager,
       SpannerResourceManager spannerResourceManager,
@@ -125,9 +204,9 @@ public abstract class SpannerToSourceDbITBase extends TemplateTestBase {
       String shardingCustomJarPath,
       String shardingCustomClassName,
       String sourceDbTimezoneOffset,
-      CustomTransformation customTransformation)
+      CustomTransformation customTransformation,
+      String sourceType)
       throws IOException {
-    // default parameters
 
     Map<String, String> params =
         new HashMap<>() {
@@ -138,13 +217,18 @@ public abstract class SpannerToSourceDbITBase extends TemplateTestBase {
             put("spannerProjectId", PROJECT);
             put("metadataDatabase", spannerMetadataResourceManager.getDatabaseId());
             put("metadataInstance", spannerMetadataResourceManager.getInstanceId());
-            put("sourceShardsFilePath", getGcsPath("input/shard.json", gcsResourceManager));
+            put(
+                "sourceShardsFilePath",
+                getGcsPath(
+                    sourceType != null ? "input/cassandra-config.conf" : "input/shard.json",
+                    gcsResourceManager));
             put("changeStreamName", "allstream");
             put("dlqGcsPubSubSubscription", subscriptionName);
             put("deadLetterQueueDirectory", getGcsPath("dlq", gcsResourceManager));
             put("maxShardConnections", "5");
             put("maxNumWorkers", "1");
             put("numWorkers", "1");
+            put("sourceType", sourceType);
           }
         };
