--- v2/spanner-to-sourcedb/src/test/resources/SpannerToCassandraSourceDbCustomTransformationIT/spanner-schema.sql
@@ -1,49 +1,0 @@
-CREATE TABLE IF NOT EXISTS alldatatypetransformation (
-    varchar_column STRING(20) NOT NULL,
-    tinyint_column INT64,
-    text_column STRING(MAX),
-    date_column DATE,
-    smallint_column INT64,
-    mediumint_column INT64,
-    int_column INT64,
-    bigint_column INT64,
-    float_column FLOAT64,
-    double_column FLOAT64,
-    decimal_column NUMERIC,
-    datetime_column TIMESTAMP,
-    timestamp_column TIMESTAMP,
-    time_column STRING(MAX),
-    year_column STRING(MAX),
-    char_column STRING(10),
-    tinytext_column STRING(MAX),
-    mediumtext_column STRING(MAX),
-    longtext_column STRING(MAX),
-    enum_column STRING(MAX),
-    bool_column BOOL,
-    other_bool_column BOOL,
-    bytes_column BYTES(MAX),
-    list_text_column JSON,
-    list_int_column JSON,
-    frozen_list_bigint_column JSON,
-    set_text_column JSON,
-    set_date_column JSON,
-    frozen_set_bool_column JSON,
-    map_text_to_int_column JSON,
-    map_date_to_text_column JSON,
-    frozen_map_int_to_bool_column JSON,
-    map_text_to_list_column JSON,
-    map_text_to_set_column JSON,
-    set_of_maps_column JSON,
-    list_of_sets_column JSON,
-    frozen_map_text_to_list_column JSON,
-    frozen_map_text_to_set_column JSON,
-    frozen_set_of_maps_column JSON,
-    frozen_list_of_sets_column JSON,
-    varint_column BYTES(MAX)
-) PRIMARY KEY(varchar_column);
-
-CREATE CHANGE STREAM allstream
-  FOR ALL OPTIONS (
-    value_capture_type = 'NEW_ROW',
-    retention_period = '7d'
-  );
