--- v2/spanner-to-sourcedb/src/test/resources/SpannerToCassandraSourceDbDatatypeIT/cassandra-config-template.conf
@@ -1,12 +1,0 @@
- # Configuration for the DataStax Java driver for Apache Cassandra®.
-  # This file is in HOCON format, see https://github.com/typesafehub/config/blob/master/HOCON.md.
-  # This file is meant to be used only in unit tests to test loading configuration from file.
-  # DO NOT USE FOR PRODUCTION.
-
-  datastax-java-driver {
-      basic.contact-points = ["##host##:##port##"]
-      basic.session-keyspace = "##keyspace##"
-    basic.load-balancing-policy {
-      local-datacenter = "datacenter1"
-    }
-  }
